requirements.txt
flask
flask-restful
flask-security-too
peewee
python-dotenv
